edge/services/data-pipeline/requirements.txt
# Data Pipeline Service - Python Dependencies

# Core Framework
fastapi==0.103.0
uvicorn[standard]==0.23.2
pydantic==2.4.0
pydantic-settings==2.0.3

# Data Processing
pandas==1.5.3
numpy==1.24.3
scipy==1.11.3

# InfluxDB Client
influxdb-client==1.38.0

# MQTT Client
paho-mqtt==1.6.1

# Monitoring
prometheus-client==0.17.1

# Kalman Filter (for data cleaning)
pykalman==0.9.5

# Signal Processing
scikit-learn==1.3.0

# Logging & Utilities
python-dotenv==1.0.0
